Moved print to external class
moved the print and main methods we got from a friend to an external
class.
